Refactoring code. devs: run `poetry install` again
